+++ scripts/make_test_data.py
@@ -4,6 +4,7 @@ NOTE: This script relies on the assumption that the user has downloaded
 `ATL08_20260118035703_05313006_007_01.h5` and placed it ../data/.
 """
 
+from collections import defaultdict
 from pathlib import Path
 
 import xarray as xr
@@ -17,6 +18,18 @@ SOURCE_DATA_PATH = (
 if not SOURCE_DATA_PATH.is_file():
     msg = "Source data path does not exist."
     raise RuntimeError(msg)
+
+
+GT_VARS_TO_INCLUDE = [
+    "canopy/h_canopy",
+    "canopy/h_canopy_uncertainty",
+    "canopy/h_median_canopy",
+    "canopy/photon_rate_can",
+    "terrain/h_te_best_fit",
+    "terrain/h_te_uncertainty",
+    "terrain/photon_rate_te",
+    "terrain/terrain_slope",
+]
 
 
 if __name__ == "__main__":
@@ -62,26 +75,29 @@ if __name__ == "__main__":
             coords="minimal",
         )
 
-        canopy_heights = ds.canopy.h_canopy
-        filtered_canopy_heights = xr.concat(
-            [
-                # First 50 observations
-                canopy_heights.isel(delta_time=slice(0, 50)),
-                # One isolated point in the middle
-                canopy_heights.isel(delta_time=int(len(canopy_heights) / 2)),
-                # Last 50 observations
-                canopy_heights.isel(delta_time=slice(-50, None)),
-            ],
-            dim="delta_time",
-            coords="minimal",
-        )
+        variables = defaultdict(dict)
+        for var_path in GT_VARS_TO_INCLUDE:
+            group_name, var_name = var_path.split("/")
+            data_var = ds[var_path]
+            variables[group_name][var_name] = xr.concat(
+                [
+                    # First 50 observations
+                    data_var.isel(delta_time=slice(0, 50)),
+                    # One isolated point in the middle
+                    data_var.isel(delta_time=int(len(data_var) / 2)),
+                    # Last 50 observations
+                    data_var.isel(delta_time=slice(-50, None)),
+                ],
+                dim="delta_time",
+                coords="minimal",
+            )
 
         test_ds = xr.DataTree.from_dict(
             {
                 f"{ground_track}/land_segments/": {
                     "latitude": filtered_lats,
                     "longitude": filtered_lons,
-                    "canopy": {"h_canopy": filtered_canopy_heights},
+                    **variables,
                 }
             },
             nested=True,