+++ tests/test_read_geom.py
@@ -18,9 +18,17 @@ def test_read_point_geoms_from_atl08(atl08_test_filepath):
     # test data - ATL08 generally has 6 ground tracks.)
     assert len(set(points.ground_track)) == 5
 
-    assert points.h_canopy is not None
-
-    assert points.delta_time is not None
+    for core_var in (
+        "h_canopy",
+        "h_canopy_uncertainty",
+        "h_median_canopy",
+        "photon_rate_can",
+        "h_te_best_fit",
+        "h_te_uncertainty",
+        "photon_rate_te",
+        "terrain_slope",
+    ):
+        assert points[core_var] is not None
 
 
 def test_lines_from_atl08_points(atl08_test_filepath):